Added option to center items horizontally
When the layout mode fitRows is set and the option fitRows.centered is set to "true", items not spanning the whole container will be centered within the container. I tried adding this functionality to the masonry layoutmode, but failed miserably. In retrospect this was pretty easy, but it took me quite some time to understand how isotope and it's layout modes work internally.

I've done some tests and am pretty sure it works, but it is possible I missed something. Hope this helps :)